Remove error if certificate cannot be found
We have had a few cases where we try to fetch a certificate before
it has been created, and cause unnecessary noise via error alerts.

If a certificate cannot be found, we now continue on.
In the case where the certificate has not been created yet, we will
fetch it on the next sync.
In the case where the certificate will never be created (due to
some unknown system issue), we will get alerted for the cluster
itself being down (as we can't scrape it, due to the missing
certificates).