controller: reconcile every minute always
Fixes https://github.com/giantswarm/giantswarm/issues/5301.

This change replaces the informer with one that in fact doesn't watch
any object. It emits ADDED event with nil object every minute
(configurable with flags). So far this operator was watching services
but didn't really read them or use them whatsoever. This were causing
problems with the reconciliation when are clusters were deleted (no
services to watch) but there were still some configuration leftovers in
the prometheus config map. This is not a perfect solution. Many things
should be redesigned here. I think it's good enough to close the PM.