controller: reconcile every minute always (#110)
Fixes giantswarm/giantswarm#5301.

This change replaces watched object with the pod the
prometheus-config-controller is running in. So far this operator was
watching services but didn't really read them or use them whatsoever.
This were causing problems with the reconciliation when are clusters
were deleted (no services to watch) but there were still some
configuration leftovers in the prometheus config map. This is not
a perfect solution. Many things should be redesigned here. I think it's
good enough to close the PM.