configmap, prometheus: return reloadThrottleError and keep finalizers
Towards https://github.com/giantswarm/giantswarm/issues/5301

There is a chance that during deleting last cluster the reload request
is throttled, the last Service is removed and the configuration is never
reloaded. Therefore finalizers are kept till the Prometheus
configuration isn't really reloaded.

Note this probably won't solve the PM as we observe that the ConfigMap
doesn't change.