configmap, prometheus: update ConfigMap when the last cluster is removed (#117)
Towards https://github.com/giantswarm/giantswarm/issues/5301

In order to do update ConfigMap when the last Service is deleted we must do the
same thing we do during update.

There is also a chance that during deleting last cluster the reload request is
throttled, the last Service is removed and the configuration is never reloaded.
Therefore finalizers are kept till the Prometheus configuration isn't really
reloaded.

NOTE this also keeps the master service up for as long as the Prometheus config isn't finally reloaded which may take up to 2 minutes as of right now. Don't know if this has some serious implications.