Fix unit tests (#127)
Verification of correct generation of scrapeconfigs requires quite big
reference data structures for expected state. In order to reduce amount of
code, only one set was explicitly defined and second one was copied & modified
from that.

This change adds proper deep copying of data when those data structures contain
pointers, slices or slices of pointers. Unit test logic as such is correct as
is the actual tested code as well. It was only the test data expectation that
got broken due to changes in Go compiler / runtime optimizations.