Scrape calico-node PODs using pod service discovery role
This change adds support for scraping PODs directly with pod service discovery
role. This way there doesn't need to be service present for
prometheus-config-controller to see but prometheus discovers the PODs directly.
This is useful e.g. for calico-node PODs which don't have nor need any service
instance.

While adapting tests for this new service discovery config, the configuration
rendering test was now converted to be so called golden file test so
expectation string doesn't need to be maintained manually.