Scrape calico-node metrics from tenant clusters (#126)
Calico provides the networking layer for Giant Swarm tenant clusters and
therefore it's important to scrape metrics from it in order to be able to
configure proper monitoring for it.

This change adds support for scraping PODs directly with pod service discovery
role. This way there doesn't need to be service present for
prometheus-config-controller to see but prometheus discovers the PODs directly.
This is useful e.g. for calico-node PODs which don't have nor need any service
instance.

While adapting tests for this new service discovery config, the configuration
rendering test was now converted to be so called golden file test so
expectation string doesn't need to be maintained manually.

Towards https://github.com/giantswarm/giantswarm/issues/6381